pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "emailinator"
version = "0.1.0"
description = "Email parser that extracts tasks using OpenAI API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "pytest>=7.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110",
    "uvicorn>=0.23",
    "requests>=2.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
